CO algorithm with HillClimbing & SimulatedAnnealing tested | remain TabuSearch

+++ gospl/src/main/java/gospl/algo/co/simannealing/ISimulatedAnnealingTransitionFunction.java
@@ -18,6 +18,6 @@ public interface ISimulatedAnnealingTransitionFunction {
 	 * @param temperature
 	 * @return
 	 */
-	public boolean getTransitionProbability(double currentEnergy, double candidateEnergy, int temperature);
+	public boolean getTransitionProbability(double currentEnergy, double candidateEnergy, double temperature);
 	
 }
